checkpoint(2)

+++ database/migrations/0001_01_01_000001_create_cache_table.php
@@ -11,17 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('cache', function (Blueprint $table) {
-            $table->string('key')->primary();
-            $table->mediumText('value');
-            $table->integer('expiration');
-        });
+        // if (!Schema::hasTable('cache')) {
+        //     Schema::create('cache', function (Blueprint $table) {
+        //         $table->string('key')->primary();
+        //         $table->mediumText('value');
+        //         $table->integer('expiration');
+        //     });
+        // }
 
-        Schema::create('cache_locks', function (Blueprint $table) {
-            $table->string('key')->primary();
-            $table->string('owner');
-            $table->integer('expiration');
-        });
+        // if (!Schema::hasTable('cache_locks')) {
+        //     Schema::create('cache_locks', function (Blueprint $table) {
+        //         $table->string('key')->primary();
+        //         $table->string('owner');
+        //         $table->integer('expiration');
+        //     });
+        // }
     }
 
     /**
@@ -29,7 +33,12 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('cache');
-        Schema::dropIfExists('cache_locks');
+        if (Schema::hasTable('cache')) {
+            Schema::dropIfExists('cache');
+        }
+
+        if (Schema::hasTable('cache_locks')) {
+            Schema::dropIfExists('cache_locks');
+        }
     }
 };
